+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/item/ItemCategorizations.kt
@@ -1,0 +1,110 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2025 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.ccbluex.liquidbounce.utils.item
+
+import net.minecraft.component.DataComponentTypes
+import net.minecraft.component.type.FoodComponent
+import net.minecraft.item.AxeItem
+import net.minecraft.item.BundleItem
+import net.minecraft.item.HoeItem
+import net.minecraft.item.ItemStack
+import net.minecraft.item.Items
+import net.minecraft.item.MiningToolItem
+import net.minecraft.item.PickaxeItem
+import net.minecraft.item.ShovelItem
+import net.minecraft.item.consume.UseAction
+import net.minecraft.registry.tag.ItemTags
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.contract
+
+/**
+ * Check if a stack is nothing (means empty slot)
+ */
+@OptIn(ExperimentalContracts::class)
+fun ItemStack?.isNothing(): Boolean {
+    contract {
+        returns(true) implies (this@isNothing != null)
+    }
+
+    return this?.isEmpty == true
+}
+
+val ItemStack.isConsumable: Boolean
+    get() = this.isFood || this.item == Items.POTION || this.item == Items.MILK_BUCKET
+
+val ItemStack.isFood: Boolean
+    get() = foodComponent != null && this.useAction == UseAction.EAT
+
+val ItemStack.foodComponent: FoodComponent?
+    get() = this.get(DataComponentTypes.FOOD)
+
+val ItemStack.isBundle
+    get() = this.isIn(ItemTags.BUNDLES)
+
+// Tools
+
+val ItemStack.isSword
+    get() = this.isIn(ItemTags.SWORDS)
+
+val ItemStack.isPickaxe
+    get() = this.isIn(ItemTags.PICKAXES)
+
+val ItemStack.isAxe
+    get() = this.isIn(ItemTags.AXES)
+
+val ItemStack.isShovel
+    get() = this.isIn(ItemTags.SHOVELS)
+
+val ItemStack.isHoe
+    get() = this.isIn(ItemTags.HOES)
+
+/**
+ * Replacement of 1.21.4 `MiningToolItem`
+ */
+val ItemStack.isMiningTool
+    get() = isAxe || isPickaxe || isShovel || isHoe
+
+// Armors
+
+val ItemStack.isFootArmor
+    get() = this.isIn(ItemTags.FOOT_ARMOR)
+
+val ItemStack.isLegArmor
+    get() = this.isIn(ItemTags.LEG_ARMOR)
+
+val ItemStack.isChestArmor
+    get() = this.isIn(ItemTags.CHEST_ARMOR)
+
+val ItemStack.isHeadArmor
+    get() = this.isIn(ItemTags.HEAD_ARMOR)
+
+val ItemStack.isPlayerArmor
+    get() = isFootArmor || isLegArmor || isChestArmor || isHeadArmor
+
+// 1.21.4 code
+
+val MiningToolItem.type: Int
+    get() = when (this) {
+        is AxeItem -> 0
+        is PickaxeItem -> 1
+        is ShovelItem -> 2
+        is HoeItem -> 3
+        else -> error("Unknown tool item $this (WTF?)")
+    }
